Rename cyber: files and update includes
